[feat]: Authorization part

+++ routers/posts.py
@@ -1,16 +1,18 @@
-from fastapi import HTTPException, status, Depends, APIRouter
+from typing import Annotated
 
+from fastapi import APIRouter, Depends, HTTPException, status
 from sqlalchemy import select
-from sqlalchemy.orm import selectinload
 from sqlalchemy.ext.asyncio import AsyncSession
+from sqlalchemy.orm import selectinload
+
 import models
+from auth import CurrentUser
+from database import get_db
 from schema import (
     PostCreate,
     PostResponse,
     PostUpdate,
 )
-from database import get_db
-from typing import Annotated
 
 router = APIRouter()
 
@@ -41,7 +43,10 @@ async def get_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
 
 @router.put("/{post_id}", response_model=PostResponse)
 async def update_post_full(
-    post_id: int, post_data: PostCreate, db: Annotated[AsyncSession, Depends(get_db)]
+    post_id: int,
+    post_data: PostCreate,
+    current_user: CurrentUser,
+    db: Annotated[AsyncSession, Depends(get_db)],
 ):
     result = await db.execute(select(models.Post).where(models.Post.id == post_id))
     post = result.scalars().first()
@@ -49,19 +54,16 @@ async def update_post_full(
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND, detail="Post not found"
         )
-    if post_data.user_id != post.user_id:
-        result = await db.execute(
-            select(models.User).where(models.User.id == post_data.user_id)
+
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to update this post.",
         )
-        user = result.scalars().first()
-        if not user:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail="User not found",
-            )
+
     post.title = post_data.title
     post.content = post_data.content
-    post.user_id = post_data.user_id
+
     await db.commit()
     await db.refresh(post, attribute_names=["author"])
 
@@ -70,7 +72,10 @@ async def update_post_full(
 
 @router.patch("/{post_id}", response_model=PostResponse)
 async def update_post_partial(
-    post_id: int, post_data: PostUpdate, db: Annotated[AsyncSession, Depends(get_db)]
+    post_id: int,
+    post_data: PostUpdate,
+    current_user: CurrentUser,
+    db: Annotated[AsyncSession, Depends(get_db)],
 ):
     result = await db.execute(
         select(models.Post)
@@ -81,6 +86,12 @@ async def update_post_partial(
     if not post:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND, detail="Post not found"
+        )
+
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to update this post.",
         )
 
     update_data = post_data.model_dump(exclude_unset=True)
@@ -95,7 +106,11 @@ async def update_post_partial(
 
 
 @router.delete("/{post_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
+async def delete_post(
+    post_id: int,
+    current_user: CurrentUser,
+    db: Annotated[AsyncSession, Depends(get_db)],
+):
     result = await db.execute(
         select(models.Post)
         .options(selectinload(models.Post.author))
@@ -105,6 +120,11 @@ async def delete_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]
     if not post:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND, detail="Post not found"
+        )
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to delete this post.",
         )
 
     await db.delete(post)
@@ -116,19 +136,15 @@ async def delete_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]
     response_model=PostResponse,
     status_code=status.HTTP_201_CREATED,
 )
-async def create_post(post: PostCreate, db: Annotated[AsyncSession, Depends(get_db)]):
-    result = await db.execute(select(models.User).where(models.User.id == post.user_id))
-    user = result.scalars().first()
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="User not found",
-        )
-
+async def create_post(
+    post: PostCreate,
+    current_user: CurrentUser,
+    db: Annotated[AsyncSession, Depends(get_db)],
+):
     new_post = models.Post(
         title=post.title,
         content=post.content,
-        user_id=post.user_id,
+        user_id=current_user.id,
     )
     db.add(new_post)
     await db.commit()
